Overhaul type handling and documentation.
This adds clear documentation regarding the use of Unicode vs. byte
strings, including which types are accepted and which types will be
output, updates all documentation to use the 'u' prefix on strings,
and updates internal use of string values to the 'u' prefix.

Notes are also added regarding the use of int vs. float types in
triplet representation conversions.